Created an AuthForm component that reduces duplicated code in login and signup forms
